pyproject.toml
[tool.poetry]
name = "docker-airflow"
version = "0.1.0"
description = ""
authors = ["Inoguchi Ryoichi <r.inoguchi@astamuse.co.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
apache-airflow = "^1.10.12"
flake8 = "^3.8.4"
autopep8 = "^1.5.4"
docker = "^4.3.1"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
